refactor: #46-코드 분리
